Fix bug that causes normcorre_batch to crash when reading from a TIFF and writing to a memory-mapped .mat file
Registered image data in mem_buffer couldn't be written to M_final.Y and M_final.Yr because the first is of class single and the latter two are of class int16. It seems that there is a line intended to cast data to be written to mem_buffer as int16, but this doesn't actually change the class of mem_buffer itself. We've addressed this for now by declaring M_final.Y and M_final.Yr as single, but I'm not sure if this has any other adverse consequences.